Adjustment to make tests public, create final task for publishing, don't skip if missing username or password
