+++ src/test/resources/gradle/buildFiles/create_debian_package_and_version_with_fileSpec.gradle
@@ -18,17 +18,21 @@ bintray {
             component = 'main'
             architecture = 'i386,noarch,amd64'
         }
-        userOrg = bintrayOrg
-        desc =  pkgDesc
+        if (!isNull(bintrayOrg)) {
+            userOrg = bintrayOrg
+        }
+        desc = pkgDesc
         labels = pkgLabels
         licenses = ['Apache-2.0']
         websiteUrl = 'https://github.com/bintray/gradle-bintray-plugin'
         issueTrackerUrl = 'https://github.com/bintray/gradle-bintray-plugin/issues'
         vcsUrl = 'https://github.com/bintray/gradle-bintray-plugin.git'
-        githubRepo = 'bintray/gradle-bintray-plugin'
-        githubReleaseNotesFile = 'README.md'
         version {
             name = versionName
         }
     }
+}
+
+boolean isNull(String value) {
+    return value == null || value == "" || value == 'null'
 }
