--- src/test/resources/gradle/buildFiles/maven_central_sync.gradle
@@ -1,20 +1,0 @@
-bintray {
-    apiUrl = bintrayApiUrl
-    user = bintrayUser
-    key = bintrayKey
-    pkg {
-        repo = mavenRepoName
-        name = pkgName
-        userOrg = bintrayOrg
-        vcsUrl = 'https://github.com/bintray/gradle-bintray-plugin.git'
-        version {
-            name = versionName
-            mavenCentralSync {
-                sync = true
-                user = mavenCentralUser
-                password = mavenCentralPassword
-                close = '1'
-            }
-        }
-    }
-}
