+++ src/test/resources/gradle/projects/configurationWithSubModules/build.gradle
@@ -1,0 +1,34 @@
+buildscript {
+    repositories {
+        mavenLocal()
+        jcenter()
+    }
+    dependencies {
+        classpath "com.jfrog.bintray.gradle:gradle-bintray-plugin:1.8.1-SNAPSHOT"
+    }
+}
+
+def rootProjectDir = projectDir
+def testsDir = new File(rootProjectDir.getParentFile().getParentFile(), 'buildFiles')
+
+def getTestFile = { fileName ->
+    testsDir.toString() + '/' + fileName
+}
+
+group = pkgName
+version = versionName
+
+allprojects {
+    repositories {
+        jcenter()
+    }
+    apply plugin: 'java'
+    apply plugin: 'maven'
+    apply plugin: 'com.jfrog.bintray'
+}
+
+apply from: getTestFile("${testName}.gradle")
+
+task wrapper(type: Wrapper) {
+    gradleVersion = '1.10'
+}
