+++ src/test/resources/gradle/projects/configurationWithSubModules/settings.gradle
@@ -1,0 +1,1 @@
+include ':api'
